Add a default value field into the stateVars conf
We better have default values. The app has states during which the
state variables don't have meaningful values so the user interface
layers have to implement fallback values anyway. It is better to
have this implemented on the logic layer so that the user
interface layers won't be bothered with this.